demo2023/biosegment/requirements.txt
audioread==2.1.9
auditok==0.2.0
h5py==3.1.0
huggingface-hub==0.4.0
kaldiio==2.17.2
librosa==0.8.0
numpy==1.19.5
omegaconf==2.0.2
pandas==1.0.1
pickleshare==0.7.5
Pillow==7.2.0
PyYAML==5.1
scikit-learn==0.22.2.post1
scipy==1.5.4
six==1.15.0
SoundFile==0.10.3.post1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
tokenizers==0.12.1
torch==1.8.1
torch-nightly==1.2.0.dev20190626
torchcontrib==0.0.2
tqdm==4.42.1
transformers==4.18.0